chore: convert tests to vitest
